Exercise 7 fix according to Dan's comments

+++ src/md/tekwill/jf4/homework/Exercise7.java
@@ -20,7 +20,7 @@ public class Exercise7 {
 
         try
         {
-            int r = Integer.parseInt(args[0]);
+            double r = Integer.parseInt(args[0]);
             double pi = Math.PI;
             double area = pi * Math.pow(r,2);
             System.out.println("Radius: " + r + System.lineSeparator() + "Area: " + area);
